# Use `FrozenDictionary` in `StaticClientStore`