--- a/src/Identity/IdentityServer/ClientStore.cs
+++ b/src/Identity/IdentityServer/ClientStore.cs
@@ -80,9 +80,9 @@ public class ClientStore : IClientStore
             return await CreateUserClientAsync(clientId);
         }
 
-        if (_staticClientStore.ApiClients.ContainsKey(clientId))
+        if (_staticClientStore.ApiClients.TryGetValue(clientId, out var client))
         {
-            return _staticClientStore.ApiClients[clientId];
+            return client;
         }
 
         return await CreateApiKeyClientAsync(clientId);