--- a/src/Identity/IdentityServer/StaticClientStore.cs
+++ b/src/Identity/IdentityServer/StaticClientStore.cs
@@ -1,4 +1,5 @@
-﻿using Bit.Core.Enums;
+﻿using System.Collections.Frozen;
+using Bit.Core.Enums;
 using Bit.Core.Settings;
 using Duende.IdentityServer.Models;
 
@@ -16,8 +17,8 @@ public class StaticClientStore
             new ApiClient(globalSettings, BitwardenClient.Desktop, 30, 1),
             new ApiClient(globalSettings, BitwardenClient.Cli, 30, 1),
             new ApiClient(globalSettings, BitwardenClient.DirectoryConnector, 30, 24)
-        }.ToDictionary(c => c.ClientId);
+        }.ToFrozenDictionary(c => c.ClientId);
     }
 
-    public IDictionary<string, Client> ApiClients { get; private set; }
+    public FrozenDictionary<string, Client> ApiClients { get; }
 }