## Type of change

<!-- (mark with an `X`) -->

```
- [ ] Bug fix
- [ ] New feature development
- [x] Tech debt (refactoring, code cleanup, dependency upgrades, etc)
- [ ] Build/deploy pipeline (DevOps)
- [ ] Other
```

## Objective
<!--Describe what the purpose of this PR is. For example: what bug you're fixing or what new feature you're adding-->
Use a `FrozenDictionary` for `StaticClientStore.ApiClients`. Weirdly, it does have a slight regression for what I call "near" misses. Where the length of the string matches a string in the dictionary but it's not in there. But the happy path is faster, as well as the "abused" client id. I think it is still worth it overall. 

```
Before
| Method      | ClientId             | Mean     | Error     | StdDev    |
|------------ |--------------------- |---------:|----------:|----------:|
| TryGetValue |                      | 2.946 ns | 0.0344 ns | 0.0305 ns |
| TryGetValue | a_muc(...)ng_up [47] | 9.237 ns | 0.0854 ns | 0.0799 ns |
| TryGetValue | connector            | 5.590 ns | 0.0369 ns | 0.0345 ns |
| TryGetValue | invalid              | 4.838 ns | 0.0318 ns | 0.0297 ns |
| TryGetValue | mobile               | 5.712 ns | 0.0518 ns | 0.0460 ns |
| TryGetValue | WEB                  | 4.285 ns | 0.0233 ns | 0.0207 ns |

After
| Method      | ClientId             | Mean      | Error     | StdDev    |
|------------ |--------------------- |----------:|----------:|----------:|
| TryGetValue |                      | 0.5561 ns | 0.0189 ns | 0.0177 ns |
| TryGetValue | a_muc(...)ng_up [47] | 0.5630 ns | 0.0153 ns | 0.0143 ns |
| TryGetValue | connector            | 1.4630 ns | 0.0105 ns | 0.0098 ns |
| TryGetValue | invalid              | 5.6536 ns | 0.0416 ns | 0.0348 ns |
| TryGetValue | mobile               | 1.4131 ns | 0.0138 ns | 0.0129 ns |
| TryGetValue | WEB                  | 4.5673 ns | 0.0523 ns | 0.0490 ns |
```


## Code changes
<!--Explain the changes you've made to each file or major component. This should help the reviewer understand your changes-->
<!--Also refer to any related changes or PRs in other repositories-->

* **perf/MicroBenchmarks/Identity/IdentityServer/StaticClientStoreTests.cs:** Benchmark
* **src/Identity/IdentityServer/StaticClientStore.cs:** Use FrozenDictionary
* **src/Identity/IdentityServer/ClientStore.cs:** Use `TryGetValue` for single dictionary access.

## Before you submit

- Please check for formatting errors (`dotnet format --verify-no-changes`) (required)
- If making database changes - make sure you also update Entity Framework queries and/or migrations
- Please add **unit tests** where it makes sense to do so (encouraged but not required)
- If this change requires a **documentation update** - notify the documentation team
- If this change has particular **deployment requirements** - notify the DevOps team